Master

See merge request telliant-mammoth/strive!431

+++ Sourcecode/Database/StriveTenant/StriveTenant/StriveCarSalon/Tables/tblLocation.sql
@@ -13,7 +13,7 @@
     [Twitter]             VARCHAR (100)      NULL,
     [Instagram]           VARCHAR (100)      NULL,
     [WifiDetail]          VARCHAR (20)       NULL,
-    [WorkhourThreshold]   INT                NULL,
+    [WorkhourThreshold]   DECIMAL (19, 4)    NULL,
     [StartTime]           DATETIMEOFFSET (7) NULL,
     [EndTime]             DATETIMEOFFSET (7) NULL,
     [IsActive]            BIT                NULL,
@@ -25,4 +25,6 @@
     CONSTRAINT [PK_tblLocation] PRIMARY KEY CLUSTERED ([LocationId] ASC),
     CONSTRAINT [FK_tblLocation_LocationType] FOREIGN KEY ([LocationType]) REFERENCES [StriveCarSalon].[tblCodeValue] ([id])
 );
+
+
